+++ src/main/java/com/mallang/mallang_backend/domain/plan/entity/Plan.java
@@ -39,7 +39,6 @@ public class Plan extends BaseTime {
     @Enumerated(EnumType.STRING)
     private PlanPeriod period; // // MONTHLY, SIX_MONTHS, YEAR
 
-    @Column(nullable = false)
     private String description;
 
     @Column(columnDefinition = "TEXT")
