+++ src/test/java/com/mallang/mallang_backend/domain/payment/service/confirm/PaymentConfirmServiceImplTest.java
@@ -3,6 +3,7 @@ package com.mallang.mallang_backend.domain.payment.service.confirm;
 import com.mallang.mallang_backend.domain.member.repository.MemberRepository;
 import com.mallang.mallang_backend.domain.payment.dto.approve.PaymentResponse;
 import com.mallang.mallang_backend.domain.payment.dto.approve.PaymentApproveRequest;
+import com.mallang.mallang_backend.domain.payment.dto.approve.Receipt;
 import com.mallang.mallang_backend.domain.payment.entity.Payment;
 import com.mallang.mallang_backend.domain.payment.repository.PaymentRepository;
 import com.mallang.mallang_backend.domain.payment.service.request.PaymentRedisService;
@@ -68,7 +69,7 @@ class PaymentConfirmServiceImplTest {
         String approvedAt = "2025-05-13T18:59:04+09:00";
         int amount = 43200;
         String method = "간편결제";
-        PaymentResponse.Receipt receipt = new PaymentResponse.Receipt("http://test.com");
+        Receipt receipt = new Receipt("http://test.com");
 
         // 반환할 가짜 응답 생성
         PaymentResponse fakeResponse = PaymentResponse.builder()
